# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mail Gateway API
  version: "1.0.0"
  description: |
    Mail Gateway REST API 스펙입니다. 이 문서는 메일 수신, 분류, 전달 관련 엔드포인트를 정의합니다.
servers:
  - url: http://localhost:3000
    description: 로컬 개발 서버
paths:
...
  /api/tenants:
    get:
      tags:
        - Tenants
      summary: 테넌트 목록
      description: 모든 테넌트를 반환합니다.
      responses:
        '200':
          description: 테넌트 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
    post:
      tags:
        - Tenants
      summary: 테넌트 생성
      description: 새 테넌트를 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - domains
              properties:
                name:
                  type: string
                domains:
                  type: string
                  description: >-
                    JSON 배열 문자열로 전달합니다. 예: ["example.com"]
                settings:
                  type: string
                  nullable: true
                  description: JSON 문자열
                enabled:
                  type: boolean
                  default: true
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tenant'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Name is required
  /api/tenants/{id}:
    get:
      tags:
        - Tenants
      summary: 테넌트 상세
      description: 테넌트 상세 정보를 반환하며 룰 개수를 포함합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 테넌트 상세
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Tenant'
                      - type: object
                        properties:
                          _count:
                            type: object
                            properties:
                              rules:
                                type: integer
        '404':
          description: 테넌트를 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Tenant not found
    put:
      tags:
        - Tenants
      summary: 테넌트 수정
      description: 테넌트 정보를 수정합니다. 모든 필드는 선택 가능합니다. settings는 null로 설정해 제거할 수 있습니다.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                domains:
                  type: string
                  description: JSON 배열 문자열
                settings:
                  type: string
                  nullable: true
                enabled:
                  type: boolean
      responses:
        '200':
          description: 수정됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tenant'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: 테넌트를 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Tenant not found
    delete:
      tags:
        - Tenants
      summary: 테넌트 삭제
      description: 테넌트를 삭제합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 삭제됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        '404':
          description: 테넌트를 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Tenant not found
  /api/rules:
    get:
      tags:
        - Rules
      summary: 분류 룰 목록
      description: 분류 룰 목록을 조회합니다.
      parameters:
        - name: tenantId
          in: query
          description: 테넌트별 필터
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 룰 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClassificationRule'
    post:
      tags:
        - Rules
      summary: 룰 생성
      description: 새 분류 룰을 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, conditions, action]
              properties:
                name:
                  type: string
                conditions:
                  type: string
                  description: JSON 배열 문자열 (RuleCondition[])
                action:
                  type: string
                  enum: [forward, log, archive, reject]
                description:
                  type: string
                  nullable: true
                priority:
                  type: integer
                  default: 0
                enabled:
                  type: boolean
                  default: true
                forwardTo:
                  type: string
                  nullable: true
                category:
                  type: string
                  nullable: true
                tenantId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ClassificationRule'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/rules/{id}:
    get:
      tags:
        - Rules
      summary: 룰 상세
      description: 특정 룰의 상세 정보를 반환합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 룰 상세
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ClassificationRule'
        '404':
          description: 룰을 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Rule not found
    put:
      tags:
        - Rules
      summary: 룰 수정
      description: 룰을 수정합니다. 모든 필드는 선택 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                conditions:
                  type: string
                  description: JSON 배열 문자열
                action:
                  type: string
                  enum: [forward, log, archive, reject]
                description:
                  type: string
                  nullable: true
                priority:
                  type: integer
                enabled:
                  type: boolean
                forwardTo:
                  type: string
                  nullable: true
                category:
                  type: string
                  nullable: true
                tenantId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        '200':
          description: 수정됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ClassificationRule'
        '404':
          description: 룰을 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Rule not found
    delete:
      tags:
        - Rules
      summary: 룰 삭제
      description: 룰을 삭제합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 삭제됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        '404':
          description: 룰을 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Rule not found
...
components:
  schemas:
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          type: string
        domains:
          type: array
          items:
            type: string
          description: 도메인 배열 (기본적으로 DB에는 JSON으로 저장)
        settings:
          type: string
          nullable: true
          description: JSON 문자열
        enabled:
          type: boolean
    ClassificationRule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: integer
          default: 0
        enabled:
          type: boolean
          default: true
        conditions:
          type: string
          description: JSON 배열 문자열 (RuleCondition[])
        action:
          type: string
          enum: [forward, log, archive, reject]
        forwardTo:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        tenantId:
          type: string
          format: uuid
          nullable: true
...
security: []
...